new location for backend, build tool changed to gradle and env tweaks in frontend
