+++ Makefile
@@ -1,0 +1,4 @@
+install:
+	gradle assemble
+	cd frontend/ && cp .env.dist .env
+	cd frontend/ && npm install
