--- backend/src/main/java/fr/openobservatory/backend/SecurityConfiguration.java
@@ -1,38 +1,0 @@
-package fr.openobservatory.backend;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configurers.CsrfConfigurer;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.CorsConfigurationSource;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-
-import java.util.List;
-
-@Configuration
-public class SecurityConfiguration {
-
-    @Bean
-    public CorsConfigurationSource corsConfigurationSource() {
-        var config = new CorsConfiguration();
-        config.setAllowedOrigins(List.of("https://poc.kevinbioj.fr"));
-        config.setAllowedHeaders(List.of("*"));
-        config.setAllowedMethods(List.of("GET"));
-        config.setAllowCredentials(false);
-        var source = new UrlBasedCorsConfigurationSource();
-        source.registerCorsConfiguration("/**", config);
-        return source;
-    }
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        return http
-                .authorizeHttpRequests(req -> req.anyRequest().permitAll())
-                .cors(Customizer.withDefaults())
-                .csrf(CsrfConfigurer::disable)
-                .build();
-    }
-}
