--- backend/src/main/java/fr/openobservatory/backend/model/celestial_body/CelestialBodyRepository.java
@@ -1,8 +1,0 @@
-package fr.openobservatory.backend.model.celestial_body;
-
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.UUID;
-
-public interface CelestialBodyRepository extends CrudRepository<CelestialBodyEntity, UUID> {
-}
