+++ build.gradle.kts
@@ -1,0 +1,25 @@
+group = "fr.openobservatory"
+version = "0.0.1-SNAPSHOT"
+description = "poc-carte-backend"
+java.sourceCompatibility = JavaVersion.VERSION_17
+
+plugins {
+    id("org.springframework.boot") version "3.0.0"
+    id("io.freefair.lombok") version "6.6.1"
+    id("com.diffplug.spotless") version "6.11.0"
+    java
+}
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation("org.springframework.boot:spring-boot-starter-data-jpa:3.0.0")
+    implementation("org.springframework.boot:spring-boot-starter-security:3.0.0")
+    implementation("org.springframework.boot:spring-boot-starter-web:3.0.0")
+
+    runtimeOnly("com.h2database:h2:2.1.214")
+
+    testImplementation("org.springframework.boot:spring-boot-starter-test:3.0.0")
+}
