+++ frontend/.env
@@ -1,0 +1,1 @@
+VITE_API_URL=http://localhost:8080
