+++ frontend/src/api/observations.ts
@@ -1,4 +1,4 @@
 export const findAllNearby = (latitude: number, longitude: number) =>
   fetch(
-    `https://pocback.kevinbioj.fr/observations/nearby?latitude=${latitude}&longitude=${longitude}`,
+    import.meta.env.VITE_API_URL + `/observations/nearby?latitude=${latitude}&longitude=${longitude}`,
   ).then((response) => response.json());
