+++ frontend/src/env.d.ts
@@ -1,0 +1,9 @@
+/// <reference types="vite/client" />
+
+interface ImportMetaEnv {
+  readonly API_URL: string
+}
+
+interface ImportMeta {
+  readonly env: ImportMetaEnv
+}
