+++ settings.gradle.kts
@@ -1,0 +1,5 @@
+/*
+ * This file was generated by the Gradle 'init' task.
+ */
+
+rootProject.name = "poc-carte"
