+++ src/main/java/fr/openobservatory/backend/model/user/UserRepository.java
@@ -1,0 +1,8 @@
+package fr.openobservatory.backend.model.user;
+
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.UUID;
+
+public interface UserRepository extends CrudRepository<UserEntity, UUID> {
+}
